backend/requirements.txt
# NexusMemo Backend Dependencies
# Python 3.11+ required

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# AI & LangChain
langchain==1.2.7
langchain-core>=1.2.7
langchain-openai==1.1.3
langchain-neo4j
langgraph>=1.0.7

# Database
sqlalchemy==2.0.36
alembic==1.14.0
aiomysql==0.2.0
pymysql==1.1.1
cryptography==44.0.0  # Required by pymysql
neo4j==5.27.0
redis==5.2.1

# Async Task Queue
celery[redis]==5.4.0

# Data Validation
pydantic==2.10.5
pydantic-settings==2.7.1

# Audio Processing
faster-whisper==1.1.1

# Utilities
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
httpx==0.28.1

# Development & Testing
pytest==8.3.4
pytest-asyncio==0.24.0
pytest-cov==6.0.0
black==24.10.0
isort==5.13.2
mypy==1.14.1
